+++ src/Bus/CommandDispatcher.php
@@ -16,17 +16,16 @@ use CloudCreativity\Modules\Infrastructure\Queue\QueueInterface;
 use CloudCreativity\Modules\Toolkit\Messages\CommandInterface;
 use CloudCreativity\Modules\Toolkit\Pipeline\MiddlewareProcessor;
 use CloudCreativity\Modules\Toolkit\Pipeline\PipeContainerInterface;
-use CloudCreativity\Modules\Toolkit\Pipeline\PipelineBuilderFactory;
-use CloudCreativity\Modules\Toolkit\Pipeline\PipelineBuilderFactoryInterface;
+use CloudCreativity\Modules\Toolkit\Pipeline\PipelineBuilder;
 use CloudCreativity\Modules\Toolkit\Result\ResultInterface;
 use RuntimeException;
 
 class CommandDispatcher implements CommandDispatcherInterface
 {
     /**
-     * @var PipelineBuilderFactoryInterface
+     * @var PipelineBuilder
      */
-    private readonly PipelineBuilderFactoryInterface $pipelineFactory;
+    private readonly PipelineBuilder $pipelineBuilder;
 
     /**
      * @var null|QueueInterface|Closure(): QueueInterface
@@ -42,15 +41,15 @@ class CommandDispatcher implements CommandDispatcherInterface
      * CommandDispatcher constructor.
      *
      * @param CommandHandlerContainerInterface $handlers
-     * @param PipelineBuilderFactoryInterface|PipeContainerInterface|null $pipeline
+     * @param PipeContainerInterface|null $middleware
      * @param null|Closure(): QueueInterface $queue
      */
     public function __construct(
         private readonly CommandHandlerContainerInterface $handlers,
-        PipelineBuilderFactoryInterface|PipeContainerInterface|null $pipeline = null,
-        Closure|null $queue = null,
+        ?PipeContainerInterface $middleware = null,
+        ?Closure $queue = null,
     ) {
-        $this->pipelineFactory = PipelineBuilderFactory::make($pipeline);
+        $this->pipelineBuilder = new PipelineBuilder($middleware);
         $this->queue = $queue;
     }
 
@@ -74,8 +73,7 @@ class CommandDispatcher implements CommandDispatcherInterface
     {
         $handler = $this->handlers->get($command::class);
 
-        $pipeline = $this->pipelineFactory
-            ->getPipelineBuilder()
+        $pipeline = $this->pipelineBuilder
             ->through([...$this->pipes, ...array_values($handler->middleware())])
             ->build(MiddlewareProcessor::wrap($handler));