+++ tests/Unit/Infrastructure/DomainEventDispatching/DeferredDispatcherTest.php
@@ -15,6 +15,7 @@ use Closure;
 use CloudCreativity\Modules\Domain\Events\DomainEventInterface;
 use CloudCreativity\Modules\Infrastructure\DomainEventDispatching\DeferredDispatcher;
 use CloudCreativity\Modules\Infrastructure\DomainEventDispatching\ListenerContainerInterface;
+use CloudCreativity\Modules\Toolkit\Pipeline\PipeContainerInterface;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
@@ -23,7 +24,12 @@ class DeferredDispatcherTest extends TestCase
     /**
      * @var ListenerContainerInterface&MockObject
      */
-    private ListenerContainerInterface $container;
+    private ListenerContainerInterface&MockObject $listeners;
+
+    /**
+     * @var MockObject&PipeContainerInterface&MockObject
+     */
+    private PipeContainerInterface&MockObject $middleware;
 
     /**
      * @var DeferredDispatcher
@@ -38,7 +44,8 @@ class DeferredDispatcherTest extends TestCase
         parent::setUp();
 
         $this->dispatcher = new DeferredDispatcher(
-            $this->container = $this->createMock(ListenerContainerInterface::class),
+            listeners: $this->listeners = $this->createMock(ListenerContainerInterface::class),
+            middleware: $this->middleware = $this->createMock(PipeContainerInterface::class),
         );
     }
 
@@ -57,7 +64,7 @@ class DeferredDispatcherTest extends TestCase
 
         $listener2Closure = static fn ($event) => $listener2->handle($event);
 
-        $this->container
+        $this->listeners
             ->expects($this->exactly(2))
             ->method('get')
             ->willReturnCallback(fn (string $name) => match ($name) {
@@ -122,7 +129,7 @@ class DeferredDispatcherTest extends TestCase
 
         $listener2Closure = static fn ($event) => $listener2->handle($event1);
 
-        $this->container
+        $this->listeners
             ->expects($this->exactly(3))
             ->method('get')
             ->willReturnCallback(fn (string $name) => match ($name) {
@@ -193,7 +200,7 @@ class DeferredDispatcherTest extends TestCase
         $listener2 = $this->createMock(TestListener::class);
         $listener3 = $this->createMock(TestListener::class);
 
-        $this->container
+        $this->listeners
             ->expects($this->exactly(3))
             ->method('get')
             ->willReturnCallback(fn (string $name) => match ($name) {
@@ -253,7 +260,7 @@ class DeferredDispatcherTest extends TestCase
         $listener3 = $this->createMock(TestListener::class);
         $listener4 = $this->createMock(TestListener::class);
 
-        $this->container
+        $this->listeners
             ->method('get')
             ->willReturnCallback(fn (string $name) => match ($name) {
                 'Listener1' => $listener1,
@@ -318,7 +325,7 @@ class DeferredDispatcherTest extends TestCase
         $listener3 = $this->createMock(TestListener::class);
         $listener4 = $this->createMock(TestListener::class);
 
-        $this->container
+        $this->listeners
             ->method('get')
             ->willReturnCallback(fn (string $name) => match ($name) {
                 'Listener1' => $listener1,
@@ -374,7 +381,7 @@ class DeferredDispatcherTest extends TestCase
     public function testNoListeners(): void
     {
         $event = $this->createMock(DomainEventInterface::class);
-        $this->container->expects($this->never())->method($this->anything());
+        $this->listeners->expects($this->never())->method($this->anything());
         $this->dispatcher->dispatch($event);
     }
 
@@ -397,6 +404,12 @@ class DeferredDispatcherTest extends TestCase
             return $next($event3);
         };
 
+        $this->middleware
+            ->expects($this->once())
+            ->method('get')
+            ->with('Middleware2')
+            ->willReturn($b);
+
         $listener = $this->createMock(TestListener::class);
 
         $listener
@@ -404,13 +417,13 @@ class DeferredDispatcherTest extends TestCase
             ->method('handle')
             ->with($this->identicalTo($event3));
 
-        $this->container
+        $this->listeners
             ->expects($this->once())
             ->method('get')
             ->with('MyListener')
             ->willReturn($listener);
 
-        $this->dispatcher->through([$a, $b]);
+        $this->dispatcher->through([$a, 'Middleware2']);
         $this->dispatcher->listen($event3::class, 'MyListener');
         $this->dispatcher->dispatch($event1);
     }
@@ -422,7 +435,7 @@ class DeferredDispatcherTest extends TestCase
     {
         $event = new TestImmediateDomainEvent();
 
-        $this->container
+        $this->listeners
             ->expects($this->once())
             ->method('get')
             ->with('Listener1')